feat(consensus): V0.2 Phase 1.5 — port u32 permille wire format
Port the wire-format change from Savitri-Testnet-V0.1.0 (commit c4ca0ab on
tmp_election2 + earlier 8707ab8) to this repo. Election candidate scores
move from f64 to u32 permille across the wire format. The signable_bytes
contract is unchanged (candidates already excluded here; the
"CRITICAL FIX v2" comment on this branch is the historical reason).

## Type changes

* savitri-lightnode/src/p2p/intra_group/mod.rs (4 sites):
  - ProposerElectionResult.candidates: Vec<(String, u32, f64)> -> Vec<(String, u32, u32)>
  - ElectionCertificate.candidates: same
  - ProposerElectionCertMessage.candidates: same
  - Local determine_proposer candidates Vec: Vec<(String, f64, u32)> -> Vec<(String, u32, u32)>
* savitri-masternode/src/proposal_validator.rs:
  - ElectionCertificate.candidates: Vec<(String, u32, f64)> -> Vec<(String, u32, u32)>
* savitri-masternode/src/libp2p_network.rs:
  - ProposerElectionCertMessage.candidates: Vec<(String, u32, f64)> -> Vec<(String, u32, u32)>

## determine_proposer cast at the wire boundary

The local f64 combined_score computation is retained for now (Phase 1.4c
canonical lookup against LatencyCanonState is NOT ported here: that
integration requires adding a `latency_canon_state` field to
IntraGroupCommunication plus subscribe/publisher wiring in
p2p/network/mod.rs, which is a larger refactor tracked separately).

f64 combined_score is cast to u32 permille just before pushing to the
candidates vec:

  let combined_permille = (combined_score * 1000.0).clamp(0.0, 1000.0) as u32;
  candidates.push((member_id.clone(), combined_permille, pou_score));

The cast is per-observer (still non-deterministic) but at least the
on-the-wire payload is integer. This unblocks future signable_bytes
re-inclusion once the canonical table convergence work lands.

## Validation

savitri-consensus compiles clean on this branch. Cluster validation was
performed on Savitri-Testnet-V0.1.0 (the integration-side branch where
Phase 1.4c is fully wired); see that repository's tmp_election2 branch.

## Scope NOT covered by this commit

- LatencyCanonState integration into IntraGroupCommunication (Phase 1.4c)
- Publisher spawn + receive handler in network/mod.rs
- Periodic DIAG[latency-canon] timer
- Re-inclusion of `candidates` and `proposer_pou_score` in signable_bytes

These are tracked as follow-up work; the foundational Phase 1 wire
format (LatencyReport, LatencyTable, publisher, state holder) shipped
in commit d4f0b24 remains available for that integration.

+++ savitri-masternode/src/proposal_validator.rs
@@ -133,7 +133,7 @@ pub struct ElectionCertificate {
     pub elected_proposer_pubkey: [u8; 32],
     pub proposer_pou_score: u32,
     pub timestamp: u64,
-    pub candidates: Vec<(String, u32, f64)>,
+    pub candidates: Vec<(String, u32, u32)>,
     pub attestations: Vec<ElectionAttestation>,
     /// First chain height at which this certificate is valid (Falla 3 anti-replay binding).
     /// `serde(default)` keeps deserialization compatible with older peers that haven't
